+++ iridauploader/tests_integration/integration_data_setup.py
@@ -1,7 +1,7 @@
-from selenium import webdriver
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
-from contextlib import contextmanager
+"""
+This file is responsible for Managing an IRIDA instance, including all the parameters associated with testing and
+database management.
+"""
 import requests
 from requests import ConnectionError
 from urllib.error import URLError
@@ -27,14 +27,7 @@ class SetupIridaData:
         self.base_url = base_url
         self.user = user
         self.password = password
-        self.driver = None
         self.branch = branch
-
-        # IRIDA info
-        self.IRIDA_PASSWORD_ID = 'password_client'
-        self.IRIDA_AUTH_CODE_ID = 'auth_code_client'
-        self.IRIDA_USER = "admin"
-        self.IRIDA_PASSWORD = "Password1!"  # new password
 
         # Database info
         self.DB_HOST = db_host
@@ -52,24 +45,30 @@ class SetupIridaData:
             '"| mysql -h ' + self.DB_HOST + ' -P ' + self.DB_PORT + ' -u test -ptest'
 
         root_dir = Path(__file__).parent.absolute()
+        sql_file = Path(root_dir, "create_sequencer_client.sql")
         output_files = Path(root_dir, "tmp", "output-files")
         reference_file = Path(root_dir, "tmp", "reference-files")
         sequence_files = Path(root_dir, "tmp", "sequence-files")
         assembly_files = Path(root_dir, "tmp", "assembly-files")
 
-        self.IRIDA_CMD = 'mvn clean compile spring-boot:start -DskipTests --quiet '\
-                         '-Dspring-boot.run.arguments=\"'\
-                         '--spring.datasource.url={} '.format(self.DB_JDBC_URL) +\
-                         '--spring.datasource.username={} '.format(self.DB_USERNAME) +\
-                         '--spring.datasource.password={} '.format(self.DB_PASSWORD) +\
-                         '--liquibase.update.database.schema=true '\
-                         '--spring.jpa.hibernate.ddl-auto= '\
-                         '--spring.jpa.properties.hibernate.hbm2ddl.import_files= '\
-                         '--sequence.file.base.directory={} '.format(sequence_files) +\
-                         '--reference.file.base.directory={} '.format(reference_file) +\
-                         '--output.file.base.directory={} '.format(output_files) +\
-                         '--assembly.file.base.directory={} '.format(assembly_files) +\
-                         '--logging.pattern.console=\"'
+        self.IRIDA_DB_UPDATE = 'mysql -h ' + self.DB_HOST + ' -P ' + self.DB_PORT + ' -u test -ptest ' + \
+                               self.DB_NAME + ' < {}'.format(sql_file)
+
+        self.IRIDA_CMD = \
+            'mvn clean compile spring-boot:start -DskipTests --quiet ' +\
+            '-Dspring-boot.run.arguments=\"' +\
+            '--spring.datasource.url={} '.format(self.DB_JDBC_URL) +\
+            '--spring.datasource.username={} '.format(self.DB_USERNAME) +\
+            '--spring.datasource.password={} '.format(self.DB_PASSWORD) +\
+            '--liquibase.update.database.schema=true ' +\
+            '--spring.jpa.hibernate.ddl-auto= ' +\
+            '--spring.jpa.properties.hibernate.hbm2ddl.import_files= ' +\
+            '--sequence.file.base.directory={} '.format(sequence_files) +\
+            '--reference.file.base.directory={} '.format(reference_file) +\
+            '--output.file.base.directory={} '.format(output_files) +\
+            '--assembly.file.base.directory={} '.format(assembly_files) +\
+            '--logging.pattern.console=' +\
+            '\"'
 
         self.IRIDA_STOP = 'mvn spring-boot:stop'
 
@@ -84,15 +83,11 @@ class SetupIridaData:
         self.REPO_PATH = path.join(self.PATH_TO_MODULE, "repos")
         self.IRIDA_PATH = path.join(self.REPO_PATH, "irida")
 
-    @contextmanager
-    def wait_for_page_load(self, timeout=30):
-        old_page = self.driver.find_element_by_tag_name('html')
-        yield
-        WebDriverWait(self.driver, timeout).until(
-            EC.staleness_of(old_page)
-        )
-
     def install_irida(self):
+        """
+        Installs IRIDA
+        :return:
+        """
         install_proc = subprocess.Popen(
             [self.INSTALL_IRIDA_EXEC, self.branch], cwd=self.PATH_TO_MODULE)
         proc_res = install_proc.wait()
@@ -100,6 +95,10 @@ class SetupIridaData:
             sys.exit(1)
 
     def reset_irida_db(self):
+        """
+        Deletes the database and creates a new empty one
+        :return:
+        """
         db_reset_proc = subprocess.Popen(self.IRIDA_DB_RESET, shell=True)
         proc_res = db_reset_proc.wait()
 
@@ -107,13 +106,33 @@ class SetupIridaData:
             print("Unable to execute:\n {cmd}".format(cmd=self.IRIDA_DB_RESET))
             sys.exit(1)
 
+    def update_irida_db(self):
+        """
+        Adds a user and a client for api operations to IRIDA database
+        :return:
+        """
+        db_update_proc = subprocess.Popen(self.IRIDA_DB_UPDATE, shell=True)
+        proc_res = db_update_proc.wait()
+
+        if proc_res == 1:  # failed to execute
+            print("Unable to execute:\n {cmd}".format(cmd=self.IRIDA_DB_UPDATE))
+            sys.exit(1)
+
     def run_irida(self):
+        """
+        Runs IRIDA with first initialization
+        Waits until IRIDA is up to return
+        :return:
+        """
         subprocess.Popen(
             self.IRIDA_CMD, cwd=self.IRIDA_PATH, shell=True)
-        self.wait_until_up()
+        self._wait_until_up()
 
-    def wait_until_up(self):
-
+    def _wait_until_up(self):
+        """
+        Queries IRIDA until it responds with HTTP 200
+        :return:
+        """
         start_time = time()
         elapsed = 0
         status_code = -1
@@ -121,62 +140,20 @@ class SetupIridaData:
 
         while status_code != HTTPStatus.OK and elapsed < self.TIMEOUT:
             try:
+                print("Trying to connect...")
                 status_code = requests.api.get(self.base_url).status_code
                 elapsed = time() - start_time
 
             except (URLError, ConnectionError):
                 sleep(10)
-
-    def start_driver(self):
-        options = webdriver.ChromeOptions()
-        options.headless = True
-        self.driver = webdriver.Chrome(options=options)
-        self.driver.implicitly_wait(30)
-
-    def login(self):
-        self.driver.get(self.base_url + "/login")
-        self.driver.find_element_by_id("loginForm_username").clear()
-        self.driver.find_element_by_id("loginForm_username").send_keys(self.user)
-        self.driver.find_element_by_id("loginForm_password").clear()
-        self.driver.find_element_by_id("loginForm_password").send_keys(self.password)
-        with self.wait_for_page_load(timeout=10):
-            self.driver.find_element_by_id("t-submit-btn").click()
-
-    def set_new_admin_pw(self):
-        self.driver.find_element_by_id("password").clear()
-        self.driver.find_element_by_id(
-            "password").send_keys(self.IRIDA_PASSWORD)
-        self.driver.find_element_by_id("confirmPassword").clear()
-        self.driver.find_element_by_id(
-            "confirmPassword").send_keys(self.IRIDA_PASSWORD)
-        with self.wait_for_page_load(timeout=100):
-            xpath = "//button[@type='submit']"
-            submit = self.driver.find_element_by_xpath(xpath)
-            submit.click()
-
-    def create_client(self):
-        self.driver.get(self.base_url + "/clients/create")
-        self.driver.find_element_by_id(
-            "clientId").send_keys(self.IRIDA_AUTH_CODE_ID)
-
-        self.driver.find_element_by_id("scope_write").click()  # for sending
-        with self.wait_for_page_load(timeout=10):
-            self.driver.find_element_by_id("create-client-submit").click()
-
-    def get_irida_secret(self):
-
-        self.driver.get(self.base_url + "/admin/clients")
-        self.driver.find_element_by_xpath(
-            "//*[contains(text(), '" + self.IRIDA_AUTH_CODE_ID + "')]").click()
-        secret = self.driver.find_element_by_id(
-            "client-secret").get_attribute("textContent")
-
-        return secret
+        print("Connected.")
 
     def stop_irida(self):
+        """
+        Stops the IRIDA mvn process
+        This will sometimes dump errors into the log, but it is harmless
+        :return:
+        """
         stopper = subprocess.Popen(
             self.IRIDA_STOP, cwd=self.IRIDA_PATH, shell=True)
         stopper.wait()
-
-    def close_driver(self):
-        self.driver.quit()