+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Changes:
 * Added GUI option to import config settings from file
   * This writes the settings loaded to the default config file s.t. the settings persist for automated uploads
 
+Developer Changes:
+* Integration tests now use SQL to add a client and user for tests, instead of selenium and ui navigation.
+
 Beta 0.7.1
 ----------
 Bug fixes: